Move API logic to MailApi
